requirements.txt
# Webserver dependencies
flask
singleton_decorator

# GCP dependencies
google-cloud-pubsub
google-cloud-storage

# PyTorch dependencies
http://download.pytorch.org/whl/cpu/torch-0.4.1-cp37-cp37m-win_amd64.whl ; sys_platform == "windows"
http://download.pytorch.org/whl/cpu/torch-0.4.1.post2-cp37-cp37m-linux_x86_64.whl ; sys_platform == "linux"
#http://download.pytorch.org/whl/cpu/torch-0.4.1-cp35-cp35m-linux_x86_64.whl ; sys_platform == "linux"
torchvision
